+++ src/index.ts
@@ -1,7 +1,7 @@
+import { Package } from "@manuth/package-json-editor";
 import { CallSite } from "callsite";
-import FileSystem = require("fs");
-import Path = require("path");
-import { isNullOrUndefined } from "util";
+import pkgUp = require("pkg-up");
+import { basename, dirname } from "upath";
 import StackTrace = require("v8-callsites");
 
 /**
@@ -10,6 +10,7 @@ import StackTrace = require("v8-callsites");
  * @param level
  * The stacktrace-level whose caller is to be determined.
  */
+// eslint-disable-next-line jsdoc/require-jsdoc
 export function GetCallerModule(level?: number): CallerModule;
 
 /**
@@ -18,6 +19,7 @@ export function GetCallerModule(level?: number): CallerModule;
  * @param method
  * The method whose caller is to be determined.
  */
+// eslint-disable-next-line jsdoc/require-jsdoc
 export function GetCallerModule(method: () => any): CallerModule;
 
 /**
@@ -28,6 +30,9 @@ export function GetCallerModule(method: () => any): CallerModule;
  *
  * @param level
  * The stacktrace-level whose caller is to be determined.
+ *
+ * @returns
+ * The caller-module.
  */
 export function GetCallerModule(method?: number | (() => any), level?: number): CallerModule
 {
@@ -41,55 +46,30 @@ export function GetCallerModule(method?: number | (() => any), level?: number): 
         origin = GetCallerModule;
         frames = method;
     }
-    else if (!isNullOrUndefined(method))
+    else if (method)
     {
         origin = method;
         frames = level;
     }
 
     stack = StackTrace(frames, origin);
-
     result = new CallerModule(stack[stack.length - 1]);
 
     /* if the caller isn't a module */
-    if (result.path === Path.basename(result.path))
+    if (result.path === basename(result.path))
     {
         result.name = result.path;
         result.root = result.callSite.isNative() ? "V8" : "node";
     }
-    /* if the caller is the topmost module */
     else
     {
-        if (result.path.split(Path.sep).indexOf("node_modules") < 0)
-        {
-            let root = Path.dirname(result.path);
-            let isModuleRoot = (fileName: string) =>
+        let packagePath = pkgUp.sync(
             {
-                let files = FileSystem.readdirSync(fileName).filter(
-                    (value, index, array) =>
-                    {
-                        return !FileSystem.lstatSync(Path.join(fileName, value)).isDirectory();
-                    });
-                return files.indexOf("package.json") > 0;
-            };
+                cwd: dirname(result.path)
+            });
 
-            while (!isModuleRoot(root))
-            {
-                root = Path.resolve(root, "..");
-            }
-
-            result.root = root;
-        }
-        /* if the caller is a submodule */
-        else
-        {
-            let pathTree = result.path.split(Path.sep);
-            let moduleFolderIndex = pathTree.indexOf("node_modules") + 1;
-
-            result.root = pathTree.slice(0, moduleFolderIndex + 1).join(Path.sep);
-        }
-
-        result.name = require(Path.join(result.root, "package.json")).name;
+        result.root = dirname(packagePath);
+        result.name = new Package(packagePath).Name;
     }
 
     return result;
@@ -136,8 +116,11 @@ export class CallerModule
 
     /**
      * Gets a string that represents the object.
+     *
+     * @returns
+     * A string that represents the object.
      */
-    public toString()
+    public toString(): string
     {
         return `${this.path}:${this.callSite.getLineNumber()}:${this.callSite.getColumnNumber()}`;
     }