+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## CallerModule [Unreleased]
+
+[Show differences](https://github.com/manuth/CallerModule/compare/v1.0.4...v1.0.5)
+
+## CallerModule v1.0.5
+### Updated
+  - The development environment
+  - The logic for determining the caller-module
+  - The module-name to `@manuth/caller-module`
+
+[Show differences](https://github.com/manuth/CallerModule/compare/v1.0.4...v1.0.5)
+
+
 ## CallerModule v1.0.4
   - Update all dependencies